+++ apps/backoffice-v2/src/lib/blocks/components/DirectorBlock/hooks/useDirectorBlock/create-directors-blocks.tsx
@@ -14,7 +14,6 @@ export const createDirectorsBlocks = ({
   revisionReasons,
   isEditable,
   isApproveDisabled,
-  isLoadingDocuments,
   workflow,
 }: Omit<ComponentProps<typeof DirectorBlock>, 'director' | 'documentSchemas'> & {
   directors: Array<ComponentProps<typeof DirectorBlock>['director']>;
@@ -26,13 +25,6 @@ export const createDirectorsBlocks = ({
   }
 
   directors?.forEach(director => {
-    const issuerCountryCode = extractCountryCodeFromDocuments(director.documents);
-    const documentSchemas = issuerCountryCode ? getDocumentsByCountry(issuerCountryCode) : [];
-
-    if (!Array.isArray(documentSchemas) || !documentSchemas.length) {
-      console.warn(`No document schema found for issuer country code of "${issuerCountryCode}".`);
-    }
-
     directorsBlocks.addCell({
       type: 'node',
       value: (
@@ -46,8 +38,6 @@ export const createDirectorsBlocks = ({
           revisionReasons={revisionReasons}
           isEditable={isEditable}
           isApproveDisabled={isApproveDisabled}
-          documentSchemas={documentSchemas}
-          isLoadingDocuments={isLoadingDocuments}
           workflow={workflow}
         />
       ),