+++ apps/backoffice-v2/src/lib/blocks/hooks/useKybRegistryInfoBlock/useKybRegistryInfoBlock.tsx
@@ -19,13 +19,11 @@ export const useKybRegistryInfoBlock = ({ pluginsOutput, workflow }) => {
           ),
         },
         workflowId: workflow?.id,
-        documents: workflow?.context?.documents,
-      } satisfies Extract<
-        Parameters<ReturnType<typeof createBlocksTyped>['addCell']>[0],
-        {
-          type: 'details';
-        }
-      >;
+        documents: workflow?.context?.documents?.map(
+          ({ details: _details, ...document }) => document,
+        ),
+        isDocumentsV2: !!workflow?.workflowDefinition?.config?.isDocumentsV2,
+      };
     }
 
     const message =