+++ packages/ui/src/components/organisms/Form/DynamicForm/fields/EntityFieldGroup/components/EntityFields/helpers/update-entities.ts
@@ -1,0 +1,10 @@
+import { IEntity } from '../../../types';
+
+export const updateEntities = (entitiesList: IEntity[], createdEntityIds: string[]) => {
+  return entitiesList.map((entity, index) => {
+    return {
+      ...entity,
+      ballerineEntityId: createdEntityIds[index],
+    };
+  });
+};