+++ packages/ui/src/components/organisms/Form/DynamicForm/fields/EntityFieldGroup/providers/EntityFieldProvider/index.ts
@@ -1,0 +1,2 @@
+export * from './EntityFieldProvider';
+export * from './hooks/external/useEntityField';