+++ packages/ui/src/components/organisms/Form/DynamicForm/fields/FieldList/FieldList.tsx
@@ -12,12 +12,11 @@ import { TDynamicFormField } from '../../types';
 import { useFieldList } from './hooks/useFieldList';
 import { StackProvider, useStack } from './providers/StackProvider';
 
-export type TFieldListValueType<T extends { _id: string }> = T[];
-
 export interface IFieldListParams {
   // jsonata expression
   defaultValue?: string;
   addButtonLabel?: string;
+  itemIndexLabel?: string;
   removeButtonLabel?: string;
 }
 
@@ -30,7 +29,11 @@ export const FieldList: TDynamicFormField<IFieldListParams> = props => {
   const { element } = props;
   const { id: fieldId, hidden } = useElement(element, stack);
   const { disabled, onFocus, onBlur } = useField(element, stack);
-  const { addButtonLabel = 'Add Item', removeButtonLabel = 'Remove' } = element.params || {};
+  const {
+    addButtonLabel = 'Add Item',
+    removeButtonLabel = 'Remove',
+    itemIndexLabel = 'Item {INDEX}',
+  } = element.params || {};
   const { items, addItem, removeItem } = useFieldList({ element });
 
   if (hidden) {
@@ -52,9 +55,14 @@ export const FieldList: TDynamicFormField<IFieldListParams> = props => {
             className="flex flex-col gap-2"
             data-testid={`${fieldId}-fieldlist-item-${index}`}
           >
-            <div className="flex flex-row justify-end">
+            <div className="flex flex-row items-center justify-between">
+              <span className="text-sm font-bold">
+                {itemIndexLabel.replace('{INDEX}', (index + 1).toString())}
+              </span>
               <span
-                className="cursor-pointer font-bold"
+                role="button"
+                tabIndex={0}
+                className="cursor-pointer text-sm"
                 onClick={() => removeItem(index)}
                 data-testid={`${fieldId}-fieldlist-item-remove-${index}`}
               >
@@ -70,8 +78,12 @@ export const FieldList: TDynamicFormField<IFieldListParams> = props => {
           </div>
         );
       })}
-      <div className="flex flex-row justify-end">
-        <Button onClick={addItem} disabled={disabled}>
+      <div className="flex flex-row justify-start">
+        <Button
+          onClick={addItem}
+          disabled={disabled}
+          className="border border-gray-200 bg-white text-[hsl(var(--muted-foreground))] shadow-[0_1px_2px_0_rgb(0_0_0_/_0.05)] hover:bg-gray-50 hover:shadow-[0_1px_2px_0_rgb(0_0_0_/_0.1)]"
+        >
           {addButtonLabel}
         </Button>
       </div>