+++ packages/ui/src/components/organisms/Form/DynamicForm/repositories/fields-repository.ts
@@ -4,6 +4,7 @@ import { CheckboxField } from '../fields/CheckboxField';
 import { CheckboxListField } from '../fields/CheckboxList';
 import { DateField } from '../fields/DateField';
 import { DOCUMENT_FIELD_TYPE, DocumentField } from '../fields/DocumentField';
+import { EntityFieldGroup } from '../fields/EntityFieldGroup';
 import { FieldList } from '../fields/FieldList';
 import { FileField } from '../fields/FileField';
 import { MultiselectField } from '../fields/MultiselectField';
@@ -23,6 +24,7 @@ export const baseFields = {
   multiselectfield: MultiselectField,
   textfield: TextField,
   fieldlist: FieldList,
+  entityfieldgroup: EntityFieldGroup,
   selectfield: SelectField,
   submitbutton: SubmitButton,
   phonefield: PhoneField,